requirements.txt
flask
pyserial
RPi.GPIO
